// src/typings/origin.d.ts
interface Song {

	user: GuildMember,
	announced: boolean,

	channel: TextBasedChannels,

	timestamp: string,
	videoDetails: {
		title: string;
		url: string;
		duration: number;
		thumbnail: string;
		author: {
		  name: string;
		};
	}

	resource: AudioResource
	
}